requirements.txt
absl-py==2.1.0
addict==2.4.0
albumentations==1.3.0
antlr4-python3-runtime==4.8
astunparse==1.6.3
audioread==3.0.1
av==14.0.1
basicsr==1.4.2
beautifulsoup4==4.12.3
bitarray==3.0.0
cachetools==5.5.1
captum==0.7.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
customtkinter==5.2.0
cycler==0.12.1
Cython==3.0.11
darkdetect==0.8.0
dataclasses==0.6
decorator==5.1.1
dlib==19.24.6
easydict==1.13
face-recognition==1.3.0
face_recognition_models==0.3.0
facexlib==0.3.0
fairseq==0.12.2
faiss-cpu==1.9.0.post1
ffmpeg-python==0.2.0
filelock==3.17.0
filterpy==1.4.5
flatbuffers==25.1.24
fonttools==4.55.6
fsspec==2024.12.0
future==1.0.0
gast==0.4.0
gdown==5.2.0
gfpgan==1.3.8
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
humanfriendly==10.0
hydra-core==1.0.7
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
insightface==0.7.3
Jinja2==3.1.5
joblib==1.4.2
keras==2.13.1
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.44.0
lmdb==1.6.2
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
networkx==3.4.2
ninja==1.11.1.3
numba==0.61.0
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.0.6
onnx==1.17.0
onnxruntime==1.15.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opennsfw2==0.10.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.1.1
praat-parselmouth==0.4.5
prettytable==3.13.0
protobuf==4.25.6
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.11
PyQt5_sip==12.16.1
pyqtgraph==0.13.7
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-vlc==3.0.21203
#pywin32==308
pyworld==0.3.5
PyYAML==6.0.2
qudida==0.0.4
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
resampy==0.4.3
rich==13.9.4
rsa==4.9
sacrebleu==2.5.1
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
sympy==1.13.1
tabulate==0.9.0
tb-nightly==2.19.0a20250126
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
#tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.1.10
timm==0.6.5
tk==0.1.0
tkinterdnd2==0.4.2
tomli==2.2.1
torch==2.0.1
torchaudio==2.0.2
torchcrepe==0.0.23
torchvision==0.15.2
tqdm==4.67.1
typing_extensions==4.5.0
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
yapf==0.43.0
